Move types to model
